scripts/nlp_case_study/requirements.txt
yt-dlp==2025.6.9
torch==2.7.1
pyannote.audio==3.3.2
openai-whisper==20240930
pydub==0.25.1
seaborn==0.13.2
plotly==6.1.2
kaleido==0.2.1
textblob==0.19.0
